airflow_docker/requirements.txt
apache-airflow-providers-mysql==4.0.1
fastf1==2.3.0